app/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
psycopg2==2.9.7
python-dotenv==1.0.0
sqlparse==0.4.4
django-crontab==0.7.1